deduplicate virtual list component
